+++ CHANGELOG.md
@@ -4,8 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: `MessagesMenu`: Add `closeButtonProps` prop.
-- Fix: `ChatLayout`: Fix error linked to `main`.
+## [11.2.0] - 2022-06-20
+
+Feature:
+
+- `MessagesMenu`: Add `closeButtonProps` prop.
+
+Fix:
+
+- `ChatLayout`: Fix error linked to `<main>`.
 
 ## [11.1.0] - 2022-06-14